% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_document.R
\name{as_digits}
\alias{as_digits}
\title{Exclude a metric from formatting}
\usage{
as_digits(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
A character string token to be processed by the knitr inline hook.
}
\description{
This function marks a metric to ensure that it is printed as digits and
not printed as text for numbers one through twelve.
}
