% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_entities.R
\name{categorize_entities}
\alias{categorize_entities}
\title{Provide a brief description and some examples to categorize entities}
\usage{
categorize_entities(
  api_key,
  entities,
  description,
  examples = NULL,
  model = "gpt-4o",
  seed = NA
)
}
\arguments{
\item{api_key}{A character string representing your OpenAI API key.}

\item{entities}{A vector of entity names.}

\item{description}{A description of the categories and the abbreviations to 
for ChatGPT to use.}

\item{examples}{A dataframe with with two columns for ChatGPT to learn from
(optional). The first column is assumed to contain entities and the 
second column is assumed to contain labels.}

\item{model}{A character string specifying the model to be used.
Defaults to "gpt-4o".}

\item{seed}{A string that is used for used to set a seed value for the random 
number generator used by the model.}
}
\value{
A dataframe with your submitted entities in the first column and 
   categorizations in the second row.
}
\description{
This function sends your vector of entity names alongside your categories
and brief descriptions to the ChatGPT API to obtain a categorized 
dataset.
}
\examples{
\dontrun{
   entities <- c("blue", "triangle", "purple", "cube")
   description <- "Categorize the entities as color or shape."
   examples <- data.frame(
    "entity" = c("blue", "triangle"),
    "label" = c("color", "shape")
   )
   categorize_entities("your_api_key_here", entities = entities, 
                       description = description, examples = examples)
   
}

}
