% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt.R
\name{chatgpt}
\alias{chatgpt}
\title{Access OpenAI Models Directly via OpenAI API}
\usage{
chatgpt(
  api_key,
  role,
  content,
  model = "gpt-3.5-turbo",
  referer = "http://localhost:3000"
)
}
\arguments{
\item{api_key}{A character string representing your OpenAI API key.}

\item{role}{A character string specifying the role for the interaction (user/system).}

\item{content}{A character string or list specifying the content for the interaction.}

\item{model}{A character string specifying the model to be used.
Defaults to "gpt-3.5-turbo".}

\item{referer}{A character string that is sent as the referer header in the request.
Set to your personal website or "http://localhost:3000" for testing purposes.}
}
\value{
The HTTP response object containing the model's response.
}
\description{
This function provides a way to directly interact with the OpenAI models using OpenAI's API.
It sets up an HTTP POST request to the OpenAI service to get a response from the specified model.
}
\examples{
\dontrun{
response <- chatgpt_xl(
  api_key = "your_api_key_here",
  role = "user",
  content = "Who will win the next Darwin award?",
  referer = "http://localhost:3000"
)
print(content(response))
}

}
