% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_entities.R
\name{consolidate_entities}
\alias{consolidate_entities}
\title{Consolidate variations in spelling for entity names using ChatGPT}
\usage{
consolidate_entities(api_key, entities, model = "gpt-4o", seed = NA)
}
\arguments{
\item{api_key}{A character string representing your OpenAI API key.}

\item{entities}{A vector of entity names.}

\item{model}{A character string specifying the model to be used.
Defaults to "gpt-4o".}

\item{seed}{A string that is used for used to set a seed value for the random 
number generator used by the model.}
}
\value{
A dataframe with the submitted strings in the first column and 
   corrections in the second column.
}
\description{
This function sends a vector of entity names along with instructions to 
the OpenAI API to consolidate variations in spelling.
}
\examples{
\dontrun{
   entities <- c("ExxonMobil", "Exxon Mobil", "ExxonMobil Corporation")
   consolidate_entities("mock_api_key", 
                        entities, 
                        model = "gpt-4o-2024-05-13", 
                        seed = 13
   )
}

}
