% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_document.R
\name{format_document}
\alias{format_document}
\title{Format document}
\usage{
format_document()
}
\description{
This function triggers a knitr inline hook which ensures that numbers are 
formatted correctly for a manuscript. Numbers from one through twelve are
spelled out unless explicitly excepted from this with the `as_digits` 
function. For large numbers, we ensure that the thousands etc. digits are 
comma separated.
}
