% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llmatch.R
\name{llmatch}
\alias{llmatch}
\title{Use ChatGPT to match a entities against a list of known relevant matches}
\usage{
llmatch(api_key, list_1, list_2, model = "gpt-4o", json_out = T, seed = NA)
}
\arguments{
\item{api_key}{A character string representing your OpenAI API key.}

\item{list_1}{description}

\item{list_2}{descr}

\item{model}{A character string specifying the model to be used.
Defaults to "gpt-4o".}

\item{json_out}{A logical value to indicate whether the response should be 
formatted as json rather than as a text string. Defaults to TRUE.}

\item{seed}{A string that is used for used to set a seed value for the 
random number generator used by the model.}
}
\value{
A dataframe with all entities from the first list and, where 
   applicable, matches from the second list.
}
\description{
Provide two lists and receive back a dataframe of fuzzy matches. This 
function sends both lists to ChatGPT wrapped in a prompt and with a role
to receive back a fuzzy-matched dataframe with all entities contained in 
the first list and matches from the second list.
}
\details{
This function provides a way to directly interact with the OpenAI models 
using OpenAI's API. It sets up an HTTP POST request to the OpenAI service 
to get a response from the specified model.
}
\examples{
\dontrun{
dataset <- c("entity a", "entity b, inc", "entity c")
of_interest <- c("entity b", "entity d")

matches <- llmatch("mock_api_key", 
                   dataset,
                   of_interest)
}
}
