% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pandoc_table.R
\name{make_pandoc_table}
\alias{make_pandoc_table}
\title{Convert a dataframe into a pandoc table}
\usage{
make_pandoc_table(df, file, max_width = 80)
}
\arguments{
\item{df}{Dataframe to be converted.}

\item{file}{Target to write resulting markdown to.}

\item{max_width}{Target line width.}
}
\description{
This function uses write_csv and pandoc to convert a dataframe into a 
table that can is usable in pandoc markdown.
}
\examples{
\dontrun{
   data(iris)
   make_pandoc_table(iris, "example.md")
} 

}
