% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt.R
\name{chatgpt}
\alias{chatgpt}
\title{Access OpenAI Models via OpenRouter API}
\usage{
chatgpt(
  api_key,
  referer = "http://localhost:3000",
  role,
  content,
  model = "gpt-3.5-turbo"
)
}
\arguments{
\item{api_key}{A character string representing your OpenAI API key.}

\item{referer}{A character string that is sent as the referer header in the request.
Set to your personal website or "http://localhost:3000" for testing purposes.
See more at https://openrouter.ai/docs.}

\item{role}{A character string specifying the role for the interaction (user/system).}

\item{content}{A character string or list specifying the content for the interaction.}

\item{model}{A character string specifying the model to be used.
Defaults to "gpt-3.5-turbo".}
}
\value{
The HTTP response object containing the model's response.
}
\description{
This function provides a way to interact with the OpenAI models using the OpenRouter API.
It sets up an HTTP POST request to the OpenRouter service to get a response from the specified model.
}
\examples{
\dontrun{
response <- chatgpt_xl(
  api_key = "your_api_key_here",
  referer = "http://localhost:3000",
  role = "user",
  content = "Who will win the next Darwin award?"
)
print(content(response))
}

}
