% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt.R
\name{chatgpt}
\alias{chatgpt}
\title{Access OpenAI Models Directly via OpenAI API}
\usage{
chatgpt(api_key, content, role = NA, model = "gpt-4o", json_out = T, seed = NA)
}
\arguments{
\item{api_key}{A character string representing your OpenAI API key.}

\item{content}{A character string or list specifying the content for the interaction.}

\item{role}{A character string specifying the role for the interaction (system). 
Defaults to "You are a helpful assistant." 
When json_out is TRUE, defaults to "You are a helpful assistant designed to output JSON."}

\item{model}{A character string specifying the model to be used.
Defaults to "gpt-4o".}

\item{json_out}{A logical value to indicate whether the response should be formatted as json rather than as a text string. 
Defaults to TRUE.}

\item{seed}{A string that is used for used to set a seed value for the random number generator used by the model.}
}
\value{
The HTTP response object containing the model's response.
}
\description{
This function provides a way to directly interact with the OpenAI models using OpenAI's API.
It sets up an HTTP POST request to the OpenAI service to get a response from the specified model.
}
\examples{
\dontrun{
response <- chatgpt(
  api_key = "your_api_key_here",
  content = "Who will win the next Darwin award?"
)
print(httr::content(response))
}

}
