% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_xl.R
\name{chatgpt_xl}
\alias{chatgpt_xl}
\title{Access Extended Array of OpenAI Models via OpenRouter}
\usage{
chatgpt_xl(api_key, referer, role, content, model = "gpt-3.5-turbo")
}
\arguments{
\item{api_key}{A character string. Your OpenAI API key.}

\item{referer}{A character string that is send as the referer. Set to your personal website or e.g. http://localhost:3000 for testing (see https://openrouter.ai/docs).}

\item{role}{A character string specifying the role for the interaction.}

\item{content}{A character string or list specifying the content for the interaction.}

\item{model}{A character string specifying the model to be used. Default is "gpt-3.5-turbo".}
}
\value{
A list containing the model's response.
}
\description{
Access Extended Array of OpenAI Models via OpenRouter
}
\examples{
\dontrun{
response <- chatgpt_xl(api_key = "your_api_key_here", role = "system", content = "Hello, ChatGPT!")
}

}
